Interviewee registration stepper
